feat: Implement JWT refresh token functionality, add global exception handling, refactor DTOs, and introduce OpenAPI configuration and data seeding.

+++ src/main/java/com/ucacue/udipsai/modules/documentos/DocumentoIdDTO.java
@@ -6,9 +6,9 @@ import lombok.Setter;
 @Getter
 @Setter
 public class DocumentoIdDTO {
-    private Long id;
+    private Integer id;
 
-    public DocumentoIdDTO(Long id) {
+    public DocumentoIdDTO(Integer id) {
         this.id = id;
     }
